HCK-14656: Move dbt helpers to another folder
