+++ forward_engineering/helpers/constraints/constants.js
@@ -1,0 +1,7 @@
+const CONSTRAINT_POSTFIX = {
+	primaryKey: 'pk',
+};
+
+module.exports = {
+	CONSTRAINT_POSTFIX,
+};
